1.收款不能为小数 2.收款列表金额单位不正确

+++ SuperService/SuperService/FacePay/FacePayResult.swift
@@ -14,10 +14,10 @@ struct FacePayResult {
   let errorCode:Int
   let waiting:Bool
   let customer:NearbyCustomer
-  let amount:Int
+  let amount:Double
   
   init(customer:NearbyCustomer,
-    amount:Int,
+    amount:Double,
     succ:Bool,
     orderNo:String?,
     errorCode:Int,
